Fix PIPE kill

+++ sky/skylet/subprocess_daemon.sh
@@ -3,11 +3,14 @@
 
 parent_pid=$1
 proc_pid=$2
+remote=${3:-0}
 
 while kill -s 0 ${parent_pid}; do sleep 1; done 
 
-# This is to avoid the PIPE outputing to the console after being killed in next line.
-pkill -PIPE -P ${proc_pid}
+if [ ${remote} -eq 0 ]; then
+    # This is to avoid the PIPE outputing to the console after being killed in next line.
+    pkill -PIPE -P ${proc_pid}
+fi
 if [[ $OSTYPE == 'darwin'* ]]; then
     # MacOS do not have the ps --forest option, it is enough to do pkill here. (pkill
     # only kills the first level descendant of the process, not recursively find all
